Allow for non GCC compilers and document
* rename IndividualIndex -> IterableBitset
* document public methods in IterableBitset
* document event_t
